requirements.txt
python-keycloak == 3.7.0
pyjwt == 2.3.0
